updates to runs on NextJS without crashing
