+++ docs/NextJS.mdx
@@ -1,0 +1,42 @@
+import { Meta, Story, Canvas } from '@storybook/blocks';
+import * as BasicStories from './demos/Basic.story';
+
+<Meta title="Docs/Getting Started/NextJS" />
+
+# NextJS
+For NextJS, you need to use the `use client` directive to make the component client side.
+If you don't, the reagraph will render only a placeholder on the server side.
+
+```jsx
+'use client';
+import { GraphCanvas } from 'reagraph';
+
+export const SimpleGraph = () => {
+  return (
+    <div className='h-[300px] w-full relative'>
+      <GraphCanvas
+        labelType='all'
+        nodes={[
+          {
+            id: 'n-1',
+            label: '1',
+          },
+          {
+            id: 'n-2',
+            label: '2',
+          },
+        ]}
+        edges={[
+          {
+            id: '1->2',
+            source: 'n-1',
+            target: 'n-2',
+            label: 'Edge 1-2',
+          },
+        ]}
+      />
+    </div>
+  );
+};
+
+```
