+++ package.json
@@ -122,7 +122,7 @@
     "typescript": "^5.3.3",
     "vite": "^5.1.4",
     "vite-plugin-checker": "^0.6.4",
-    "vite-plugin-css-injected-by-js": "^3.4.0",
+    "vite-plugin-css-injected-by-js": "^3.5.2",
     "vite-plugin-dts": "^3.7.3",
     "vite-plugin-svgr": "^4.2.0",
     "vite-tsconfig-paths": "^4.3.1",
