+++ src/CameraControls/index.ts
@@ -1,4 +1,13 @@
-export * from './CameraControls';
-export * from './useCameraControls';
-export * from './useCenterGraph';
+export { CameraControls } from './CameraControlsWrapper';
+export type {
+  CameraControlsProps,
+  CameraControlsRef,
+  CameraMode
+} from './CameraControls';
+export type {
+  CenterGraphInput,
+  CenterGraphOutput,
+  CenterNodesParams,
+  FitNodesParams
+} from './useCenterGraph';
 export * from './utils';
