+++ src/GraphCanvas/index.ts
@@ -1,0 +1,2 @@
+export { GraphCanvas } from './GraphCanvas';
+export type { GraphCanvasProps, GraphCanvasRef } from './GraphCanvasContent';
