+++ src/GraphScene/index.ts
@@ -1,0 +1,2 @@
+export { GraphScene } from './GraphScene';
+export type { GraphSceneProps, GraphSceneRef } from './GraphSceneContent';
