+++ src/index.ts
@@ -3,11 +3,9 @@ export * from './GraphScene';
 export * from './types';
 export * from './utils';
 export * from './layout';
-export * from './symbols';
 export * from './sizing';
 export * from './selection';
 export * from './collapse';
 export * from './RadialMenu';
 export * from './CameraControls';
-export * from './useGraph';
 export * from './themes';
