+++ src/selection/index.ts
@@ -1,3 +1,4 @@
-export * from './Lasso';
+export { Lasso } from './Lasso';
+export type { LassoProps, LassoType } from './LassoContent';
 export * from './useSelection';
 export * from './utils';
