+++ src/store.ts
@@ -1,5 +1,6 @@
-import { StoreApi, create } from 'zustand';
-import createContext from 'zustand/context';
+import { createContext, useContext, FC, ReactNode } from 'react';
+import React from 'react';
+import { StoreApi, create, useStore as useZustandStore } from 'zustand';
 import {
   InternalGraphEdge,
   InternalGraphNode,
@@ -55,8 +56,7 @@ export interface GraphState {
   setClusterPosition: (id: string, position: CenterPositionVector) => void;
 }
 
-export const { Provider, useStore } = createContext<StoreApi<GraphState>>();
-
+// Create a store factory function
 export const createStore = ({
   actives = [],
   selections = [],
@@ -67,10 +67,10 @@ export const createStore = ({
     theme: {
       ...theme,
       edge: {
-        ...theme.edge,
+        ...theme?.edge,
         label: {
-          ...theme.edge.label,
-          fontSize: theme.edge.label.fontSize ?? 6
+          ...theme?.edge?.label,
+          fontSize: theme?.edge?.label?.fontSize ?? 6
         }
       }
     },
@@ -206,3 +206,25 @@ export const createStore = ({
         return state;
       })
   }));
+
+const isServer = typeof window === 'undefined';
+const defaultStore = createStore({});
+const StoreContext = isServer
+  ? null
+  : createContext<StoreApi<GraphState>>(defaultStore);
+
+export const Provider: FC<{
+  children: ReactNode;
+  store?: StoreApi<GraphState>;
+}> = ({ children, store = defaultStore }) => {
+  if (isServer) {
+    return children;
+  }
+
+  return React.createElement(StoreContext.Provider, { value: store }, children);
+};
+
+export const useStore = <T>(selector: (state: GraphState) => T): T => {
+  const store = useContext(StoreContext);
+  return useZustandStore(store, selector);
+};
