+++ src/symbols/Edge.tsx
@@ -16,9 +16,9 @@ import { Line } from './Line';
 import { useStore } from '../store';
 import { ContextMenuEvent, InternalGraphEdge } from '../types';
 import { Html, useCursor } from '@react-three/drei';
-import { useHoverIntent } from '../utils/useHoverIntent';
+import { useHoverIntent } from '../hooks';
 import { Euler, Vector3 } from 'three';
-import { ThreeEvent } from '@react-three/fiber';
+import type { ThreeEvent } from '@react-three/fiber';
 
 /**
  * Label positions relatively edge.
