+++ src/symbols/Node.tsx
@@ -20,12 +20,11 @@ import {
   CollapseProps
 } from '../types';
 import { Html, useCursor } from '@react-three/drei';
-import { useCameraControls } from '../CameraControls';
+import { useCameraControls } from '../CameraControls/useCameraControls';
 import { useStore } from '../store';
-import { useDrag } from '../utils/useDrag';
+import { useDrag, useHoverIntent } from '../hooks';
 import { Icon } from './nodes';
-import { useHoverIntent } from '../utils/useHoverIntent';
-import { ThreeEvent } from '@react-three/fiber';
+import type { ThreeEvent } from '@react-three/fiber';
 
 export interface NodeProps {
   /**
