+++ src/utils/index.ts
@@ -5,6 +5,4 @@ export * from './position';
 export * from './layout';
 export * from './graph';
 export * from './cluster';
-export * from './useHoverIntent';
-export * from './useDrag';
 export * from './paths';
